Languages dashboard with MySQL

+++ Java/Spring/languages/src/main/java/com/zbay/languages/controllers/LanguageController.java
@@ -40,28 +40,41 @@ public class LanguageController {
     @RequestMapping(path="/languages/{index}", method=RequestMethod.GET)
     public String language(Model model, @PathVariable("index") int index) {
         List<Language> languages = languageService.allLanguages();
-        model.addAttribute("index", index);
-        model.addAttribute("language", languages.get(index));
-        return "language";
+    	if(index < languages.size()) {
+            model.addAttribute("index", index);
+            model.addAttribute("language", languages.get(index));
+            return "language";
+    	}
+    	return "redirect:/languages";
     }
     
     @RequestMapping(path="/languages/delete/{index}", method=RequestMethod.GET)
     public String deleteLanguage(@PathVariable("index") int index) {
-    	this.languageService.deleteLangByIndex(index);
+    	List<Language> languages = languageService.allLanguages();
+    	if(index < languages.size()) {
+    		this.languageService.deleteLangByIndex(index);	
+    	}
     	return "redirect:/languages";
     }
     
     @RequestMapping(path="/languages/edit/{index}", method=RequestMethod.GET)
     public String editLanguage(Model model, @PathVariable("index") int index) {
         List<Language> languages = languageService.allLanguages();
-        model.addAttribute("index", index);
-        model.addAttribute("language", languages.get(index));
-        return "edit";
+        if(index < languages.size()) {
+            model.addAttribute("index", index);
+            model.addAttribute("language", languages.get(index));
+            return "edit";	
+        }
+        return "languages";
     }
     
     @PostMapping("/languages/edit/{index}")
     public String editLanguageSubmit(@PathVariable("index") int index, @Valid @ModelAttribute("language") Language language, BindingResult result) {
-        if (result.hasErrors()) {
+    	List<Language> languages = languageService.allLanguages();
+    	if(index < languages.size()) {
+        	return "redirect:/languages";
+        }
+    	if (result.hasErrors()) {
             return "edit";
         }
         else {
